refactor(problem4): 클래스를 역할 별로 하위 패키지에 분류
