+++ src/main/java/onboarding/Problem4.java
@@ -1,7 +1,6 @@
 package onboarding;
 
-import onboarding.problem4.BlueFrogService;
-
+import onboarding.problem4.service.BlueFrogService;
 
 public class Problem4 {
     public static String solution(String word) {
